Added code to set up GDT
GDT is a necessary prerequisite to implement interrupts.
GRUB may provide a valid GDTR if you are lucky but it is
not guaranteed as per the Multiboot standard.

We do not use segmentation for memory protection. We merely
use it to get a 4GB flat memory space. Therefore, the CS and
DS are overlapped.